requirements.txt
google-genai
python-dotenv
chromadb
streamlit
pdfplumber
python-docx
tqdm
